+++ src/main.js
@@ -2,6 +2,8 @@ import "./style.css";
 import * as HIIcons from "oh-vue-icons/icons/hi";
 import * as RIcons from "oh-vue-icons/icons/ri";
 
+import 'leaflet/dist/leaflet.css';
+
 import { OhVueIcon, addIcons } from "oh-vue-icons";
 
 import App from "./App.vue";
