+++ src/views/pages/LoginPage.vue
@@ -4,10 +4,10 @@
   </SingleSideBar>
   <div v-else
        :style="`background-color: ${appAccentColor}`"
-       :class="`h-screen flex items-center  justify-center max-h-screen p-8 md:p-20`"
+       :class="`h-screen flex items-center  justify-center max-h-screen  md:p-20`"
   >
     <div
-        class="flex w-full md:w-3/4 h-fit md:h-full bg-white shadow rounded-lg overflow-clip"
+        class="flex w-full h-full md:w-3/4 md:h-full bg-white md:rounded-lg overflow-clip"
     >
       <div class="items-center flex w-full md:w-1/2 p-8 md:p-10 h-full">
         <LoginForm/>
